test: split cart utils into dedicated fixture
